Replaced 'int' for 'size_t' where approprate.  And written lots of tests for Array class and fixed the bugs uncovered by the tests.

+++ src/Device.cpp
@@ -23,7 +23,7 @@ Device::~Device() {
     }
 }
 
-void Device::getReadingFromCSV(const char * filename, const int samplePeriod)
+void Device::getReadingFromCSV(const char * filename, const size_t samplePeriod)
 {
     Signature * sig = new Signature( filename, samplePeriod );
     signatures.push_back( sig );
