+++ src/Statistic.h
@@ -29,7 +29,7 @@ struct Statistic {
     /************************
      *  Member functions    *
      ************************/
-    Statistic(const Array<T>& data, const int beginning=0, int end=0)
+    Statistic(const Array<T>& data, const size_t beginning=0, size_t end=0)
     : mean(0), stdev(0)
     {
         register T accumulator = 0;
@@ -42,13 +42,13 @@ struct Statistic {
         assert(end >  beginning);
         assert(end <= data.size);
 
-        int length = end-beginning;
+        size_t length = end-beginning;
 
         // Initialise min and max to the first value
         min = max = data[beginning];
 
         // Find the mean, min and max
-        for (int i=beginning; i<end; i++) {
+        for (size_t i=beginning; i<end; i++) {
             currentVal = data[i];
 
             accumulator += currentVal;
@@ -63,7 +63,7 @@ struct Statistic {
 
         // Find the population standard deviation
         accumulator = 0;
-        for (int i=beginning; i<end; i++) {
+        for (size_t i=beginning; i<end; i++) {
             accumulator += pow( ( data[i] - mean ), 2 );
         }
         stdev = sqrt(accumulator / length);
